Move search API functions to server directory and remove route handler
